% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prd_pickMtrl2.R
\name{prd_pickMtrl2Server}
\alias{prd_pickMtrl2Server}
\title{处理逻辑}
\usage{
prd_pickMtrl2Server(input, output, session, dms_token, erp_token, apsToken)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
prd_pickMtrl2Server()
}
