% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_outStock.R
\name{sal_outStockServer}
\alias{sal_outStockServer}
\title{处理逻辑}
\usage{
sal_outStockServer(input, output, session, dms_token, erp_token, apsToken)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
sal_outStockServer()
}
