% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_returnStock.R
\name{sal_returnStockSyncServer}
\alias{sal_returnStockSyncServer}
\title{处理逻辑}
\usage{
sal_returnStockSyncServer(
  input,
  output,
  session,
  dms_token,
  erp_token,
  apsToken
)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}

\item{erp_token}{口令}

\item{apsToken}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
sal_returnStockSyncServer()
}
