% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_AssembledApp.R
\name{stk_AssembledAppSelectServer}
\alias{stk_AssembledAppSelectServer}
\title{处理逻辑}
\usage{
stk_AssembledAppSelectServer(input, output, session, dms_token, erp_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
stk_AssembledAppSelectServer()
}
