% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_transferDirect.R
\name{stk_transferDirectSyncServer}
\alias{stk_transferDirectSyncServer}
\title{处理逻辑}
\usage{
stk_transferDirectSyncServer(
  input,
  output,
  session,
  dms_token,
  erp_token,
  apsToken
)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}

\item{erp_token}{口令}

\item{apsToken}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
stk_transferDirectSyncServer()
}
