% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_pickMtrl.R
\name{sub_pickMtrlServer}
\alias{sub_pickMtrlServer}
\title{处理逻辑}
\usage{
sub_pickMtrlServer(input, output, session, dms_token, erp_token, apsToken)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
sub_pickMtrlServer()
}
